feat(DTFS2-7795): separate out view model creation
