+++ gef-ui/templates/partials/amendments/check-your-answers.njk
@@ -42,7 +42,7 @@
 
         <input type="hidden" name="_csrf" value="{{ csrfToken }}">
 
-        {{ amendmentSummaryList.render(amendmentRows, eligibilityRows, effectiveDateRows) }}
+        {{ amendmentSummaryList.render(amendmentSummaryListParams) }}
 
         <div class="govuk-button-group">
           {{ govukButton({
